requirements.txt
Flask>=0.12.2,<0.13
flask-cors>=3.0,<4.0
flask-compress>=1.4,<1.5
flask-restplus>=0.9.2,<1.0

msgpack-python>=0.4.8,<0.5.0
ujson>=1.35,<2.0
csirtg_indicator<=1.0

setuptools>=36
cython>=0.2
pyzmq>=16.0.2,<17.0
maxminddb>=1.2.0,<1.3
geoip2>=2.2.0,<2.3
pygeoip==0.3.2
dnspython>=1.15.0,<1.16
SQLAlchemy>=1.2,<1.3
networkx>=4.1,<4.2

csirtg_smrt>=0.0.0b7,<1.0
csirtg_dnsdb==0.0.0a4

